+++ src/Zcrudgen.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace ZaidYasyaf\Zcrudgen;
+
+class Zcrudgen
+{
+}